Marrying up the tests to the controllers

+++ database/migrations/2024_02_25_234228_create_components_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('components', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('component_type_id');
-            $table->unsignedInteger('turbine_id');
+            $table->unsignedBigInteger('component_type_id');
+            $table->unsignedBigInteger('turbine_id');
             $table->foreign('turbine_id')->references('id')->on('turbines');
             $table->foreign('component_type_id')->references('id')->on('component_types');
             $table->softDeletes();
