+++ database/migrations/2024_02_25_234233_create_grades_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('grades', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('inspection_id');
+            $table->unsignedBigInteger('inspection_id');
             $table->unsignedBigInteger('component_id');
             $table->unsignedBigInteger('grade_type_id');
             $table->foreign('inspection_id')->references('id')->on('inspections');
